+++ src/main/java/com/example/demo/study/implement/StudyReader.java
@@ -6,6 +6,10 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.example.demo.dto.SliceResponse;
+import com.example.demo.dto.study.CafeStudySearchListRequest;
+import com.example.demo.exception.ExceptionType;
+import com.example.demo.study.infrastructure.CafeStudyQueryRepository;
 import org.springframework.stereotype.Component;
 
 import com.example.demo.exception.CafegoryException;
@@ -22,6 +26,7 @@ public class StudyReader {
 
 	private final CafeStudyRepository cafeStudyRepository;
 	private final StudyMemberReader studyMemberReader;
+	private final CafeStudyQueryRepository cafeStudyQueryRepository;
 
 	public Study read(Long cafeStudyId) {
 		CafeStudyEntity cafeStudyEntity = cafeStudyRepository.findById(cafeStudyId)
@@ -31,12 +36,25 @@ public class StudyReader {
 	}
 
 	public List<Study> readUpcomingBy(Long memberId, LocalDateTime now) {
-		List<Participant> upcomings = studyMemberReader.read(memberId);
+		List<Participant> upcomings = studyMemberReader.readMyUpcomingsBy(memberId);
 		List<Long> studyIds = upcomings.stream().map(Participant::getStudyId).collect(Collectors.toList());
 
 		return cafeStudyRepository.findUpcomingsBy(studyIds, now).stream()
 			.map(CafeStudyEntity::toStudy)
 			.collect(Collectors.toList());
+	}
+
+	public SliceResponse<CafeStudyEntity> searchCafeStudies(CafeStudySearchListRequest request) {
+		return cafeStudyQueryRepository.findCafeStudies(request);
+	}
+
+	public List<CafeStudyEntity> readAllWithCoordinatorBy(Long cafeId) {
+		return cafeStudyRepository.findAllByCafeId(cafeId);
+	}
+
+	public CafeStudyEntity readStudyEntity(Long cafeStudyId) {
+		return cafeStudyRepository.findById(cafeStudyId)
+			.orElseThrow(() -> new CafegoryException(ExceptionType.CAFE_STUDY_NOT_FOUND));
 	}
 
 }
