+++ src/main/java/com/example/demo/study/service/CafeStudyQueryService.java
@@ -2,6 +2,7 @@ package com.example.demo.study.service;
 
 import java.util.List;
 
+import com.example.demo.study.implement.StudyReader;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,7 +12,6 @@ import com.example.demo.dto.study.CafeStudySearchListRequest;
 import com.example.demo.dto.study.CafeStudySearchListResponse;
 import com.example.demo.implement.study.CafeStudyCommentEntity;
 import com.example.demo.implement.study.CafeStudyCommentReader;
-import com.example.demo.study.implement.CafeStudyReader;
 import com.example.demo.study.infrastructure.CafeStudyEntity;
 
 import lombok.RequiredArgsConstructor;
@@ -21,7 +21,7 @@ import lombok.RequiredArgsConstructor;
 @Transactional(readOnly = true)
 public class CafeStudyQueryService {
 
-	private final CafeStudyReader cafeStudyReader;
+	private final StudyReader cafeStudyReader;
 	private final CafeStudyCommentReader cafeStudyCommentReader;
 
 	public SliceResponse<CafeStudySearchListResponse> searchCafeStudiesByDynamicFilter(
@@ -31,7 +31,7 @@ public class CafeStudyQueryService {
 	}
 
 	public CafeStudyDetailResponse getCafeStudyDetail(Long cafeStudyId) {
-		CafeStudyEntity cafeStudy = cafeStudyReader.read(cafeStudyId);
+		CafeStudyEntity cafeStudy = cafeStudyReader.readStudyEntity(cafeStudyId);
 		List<CafeStudyCommentEntity> comments = cafeStudyCommentReader.readAllBy(cafeStudyId);
 
 		return CafeStudyDetailResponse.of(cafeStudy, comments);
